+++ CHANGELOG.md
@@ -5,9 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2021-08-04
+### Updated
+- Updated StateChange package to 0.3.0
 
-## [Unreleased]
-- Restructuring packages for TypeScript output
+## [0.5.0] - 2021-08-03 
+### Updated
+- Updated StateChange package to 0.2.0
+
+## [0.4.0] - 2021-08-02
+### Updated
+- Updated Event package to 0.3.0
+
+
+## [0.3.0] - 2021-07-23
+### Added
+- Renovate
+### Updated
+- Updated linkProp package to 0.2.0
+- Restructured packages for TypeScript output
+
 
 ## [0.2.0]
 ### Changed